Fixed things like indenting, overriding of parent's methods. Added diskStats class.

+++ buseRAMDevice.cpp
@@ -20,16 +20,16 @@ namespace buse {
 		free(data);
 	}
 
-	uint32_t buseRAMDevice::read(void* buf, uint32_t len, uint64_t offset) {
+	int buseRAMDevice::read(void* buf, size_t len, off64_t offset) {
 		buseOperations::read(buf, len, offset);
-		if(len + offset > this->getSize()) return EFBIG;
+		if(len + (uint64_t)offset > this->getSize()) return EFBIG;
 		memcpy(buf, (char *) data + offset, len);
 		return 0;
 	}
 
-	uint32_t buseRAMDevice::write(const void* buf, uint32_t len, uint64_t offset) {
+	int buseRAMDevice::write(const void* buf, size_t len, off64_t offset) {
 		buseOperations::write(buf, len, offset);
-		if(len + offset > this->getSize()) return EFBIG;
+		if(len + (uint64_t)offset > this->getSize()) return EFBIG;
 		memcpy((char *) data + offset, buf, len);
 		return 0;
 	}
